refactor(firestore): move auto-id writes into engine
